option to specify the jenkins slave user
Prior to this commit, the Jenkins slaves would run under the same user
that started the Mesos master. This is a default of the Jenkins Mesos
plugin. But this meant that if Marathon is running as root, all our
builds would also take place as root.

Fortunately, the Mesos plugin allows us to specify the user that the
slave will run under.

This commit adds the '--user' argument, allowing us to specify a
non-root user to run the Jenkins slave under on each of the Mesos
slaves. Note that this user must already be present on the system.